requirements.txt
python-mailcow
python-dotenv
requests
